Add files via upload
The "Recognizing Parking Spaces" first identifies the maximum umber of vehicles parked in a parking lot. It then saves their coordinates into a list.

In the Testing, a new frame of the parking lot is imported and then empty spaces on this new frame are marked with red whiile the frames containing a car are marked Green.

Further:
I intend to include a navigational system which could direct an upcoming vehicle towards an empty parking space.